app/requirements.txt
flask
opencv-python-headless
google-cloud-vision
google-cloud-storage
numpy
yt-dlp